tests/test_AI.py
import pytest
from AI import AI  # Assuming the AI class is in 'your_module.py'

class MockBoard:
    def __init__(self):
        self.board = [[None for _ in range(6)] for _ in range(7)]
        self.rows = 6
        self.columns = 7

    def available_moves(self):
        return [i for i in range(self.columns)]

    def make_move(self, player, col):
        for row in range(self.rows):
            if self.board[col][row] is None:
                self.board[col][row] = player
                return True
        return False

    def get_winner(self):
        return None  # Simplified for this mock version

@pytest.fixture
def ai():
    return AI(1)

@pytest.fixture
def board():
    return MockBoard()

def test_initialization(ai):
    assert ai.plyr_num == 1

def test_set_player_number(ai):
    ai.set_player_number(2)
    assert ai.plyr_num == 2

def test_minimax(ai, board):
    score = ai.minimax(board, 1, -float('inf'), float('inf'), True)
    assert isinstance(score, int)

def test_evaluate_board(ai, board):
    score = ai.evaluate_board(board)
    assert isinstance(score, int)

# Additional tests can be added for other methods and more complex scenarios.
